feat(INJI-473) - Move logState to commonUtil to remove cyclic dependency.
