+++ shared/commonUtil.ts
@@ -1,5 +1,7 @@
 import {Platform} from 'react-native';
 import argon2 from 'react-native-argon2';
+import {AnyState} from 'xstate';
+import {getDeviceNameSync} from 'react-native-device-info';
 
 export const hashData = async (
   data: string,
@@ -27,3 +29,25 @@ export default function testIDProps(id) {
 export const removeWhiteSpace = (str: string) => {
   return str.replace(/\s/g, '');
 };
+
+export function logState(state: AnyState) {
+  const data = JSON.stringify(
+    state.event,
+    (key, value) => {
+      if (key === 'type') return undefined;
+      if (typeof value === 'string' && value.length >= 100) {
+        return value.slice(0, 100) + '...';
+      }
+      return value;
+    },
+    2,
+  );
+  console.log(
+    `[${getDeviceNameSync()}] ${state.machine.id}: ${
+      state.event.type
+    } -> ${state.toStrings().pop()}\n${
+      data.length > 300 ? data.slice(0, 300) + '...' : data
+    }
+    `,
+  );
+}
